requirements.txt
pytest-playwright==0.7.0
requests==2.26.0
args==0.1.0
discord==1.0.1
discord.py==1.7.2
